Relocate GReaderColumn into a separate file and rename into dt::read::Column
