+++ Makefile
@@ -443,6 +443,7 @@ fast_objects = $(addprefix $(BUILDDIR)/, \
 	python/list.o             \
 	python/long.o             \
 	python/string.o           \
+	read/column.o             \
 	rowindex.o                \
 	rowindex_array.o          \
 	rowindex_slice.o          \
@@ -629,7 +630,7 @@ $(BUILDDIR)/csv/py_csv.h: c/csv/py_csv.h $(BUILDDIR)/py_utils.h
 	@echo • Refreshing c/csv/py_csv.h
 	@cp c/csv/py_csv.h $@
 
-$(BUILDDIR)/csv/reader.h: c/csv/reader.h $(BUILDDIR)/column.h $(BUILDDIR)/datatable.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/array.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/utils/shared_mutex.h $(BUILDDIR)/writebuf.h
+$(BUILDDIR)/csv/reader.h: c/csv/reader.h $(BUILDDIR)/column.h $(BUILDDIR)/read/column.h $(BUILDDIR)/datatable.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/array.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/utils/shared_mutex.h $(BUILDDIR)/writebuf.h
 	@echo • Refreshing c/csv/reader.h
 	@cp c/csv/reader.h $@
 
@@ -670,6 +671,11 @@ $(BUILDDIR)/python/long.h: c/python/long.h $(BUILDDIR)/utils/pyobj.h
 $(BUILDDIR)/python/string.h: c/python/string.h $(BUILDDIR)/utils/pyobj.h
 	@echo • Refreshing c/python/string.h
 	@cp c/python/string.h $@
+
+
+$(BUILDDIR)/read/column.h: c/read/column.h $(BUILDDIR)/memrange.h $(BUILDDIR)/utils/pyobj.h $(BUILDDIR)/writebuf.h
+	@echo • Refreshing c/read/column.h
+	@cp c/read/column.h $@
 
 
 $(BUILDDIR)/utils/alloc.h: c/utils/alloc.h
@@ -907,6 +913,10 @@ $(BUILDDIR)/python/long.o : c/python/long.cc $(BUILDDIR)/python/long.h $(BUILDDI
 	@$(CC) -c $< $(CCFLAGS) -o $@
 
 $(BUILDDIR)/python/string.o : c/python/string.cc $(BUILDDIR)/python/string.h $(BUILDDIR)/utils/exceptions.h
+	@echo • Compiling $<
+	@$(CC) -c $< $(CCFLAGS) -o $@
+
+$(BUILDDIR)/read/column.o : c/read/column.cc $(BUILDDIR)/read/column.h $(BUILDDIR)/csv/reader.h $(BUILDDIR)/csv/reader_parsers.h $(BUILDDIR)/python/string.h
 	@echo • Compiling $<
 	@$(CC) -c $< $(CCFLAGS) -o $@
