+++ c/csv/reader_arff.h
@@ -25,7 +25,7 @@ class ArffReader
   const char* ch;  // pointer to the current reading location
   int line;        // current line number within the input (1-based)
   int : 32;
-  std::vector<GReaderColumn> columns;
+  std::vector<dt::read::Column> columns;
 
 public:
   ArffReader(GenericReader&);
